Added units of measure retrieved method and refactored adapter
